+++ openeo/rest/job.py
@@ -2,16 +2,13 @@ import datetime
 import logging
 import time
 import typing
-import urllib.request
 from pathlib import Path
-from typing import List, Union, Dict
+from typing import List, Union, Dict, Optional
 
-from deprecated import deprecated
-from requests import ConnectionError
+from deprecated.sphinx import deprecated
+from requests import ConnectionError, Response
 
-from openeo.job import Job, JobResult, JobLogEntry
 from openeo.rest import OpenEoClientException, JobFailedException
-from openeo.rest.result import Result
 from openeo.util import ensure_dir
 
 if hasattr(typing, 'TYPE_CHECKING') and typing.TYPE_CHECKING:
@@ -22,24 +19,36 @@ if hasattr(typing, 'TYPE_CHECKING') and typing.TYPE_CHECKING:
 logger = logging.getLogger(__name__)
 
 
-class RESTJobResult(JobResult):
-    def __init__(self, url):
-        self.url = url
+class JobLogEntry:
+    """
+    A Job log entry.
+    """
 
-    def save_as(self, target_file):
-        urllib.request.urlretrieve(self.url, target_file)
+    def __init__(self, log_id: str, level: str, message: str):
+        self.log_id = log_id
+        self.level = level
+        self.message = message
 
 
-class RESTJob(Job):
+class RESTJob:
+    """
+    Handle for an openEO batch job, allowing it to describe, start, cancel, inspect results, etc.
+    """
+    # TODO: rename this to BatchJob?
 
     def __init__(self, job_id: str, connection: 'Connection'):
-        super().__init__(job_id)
+        self.job_id = job_id
+        """Unique identifier of the batch job (string)."""
+
         self.connection = connection
+
+    def __repr__(self):
+        return '<{c} job_id={i!r}>'.format(c=self.__class__.__name__, i=self.job_id)
 
     def describe_job(self):
         """ Get all job information."""
         # GET /jobs/{job_id}
-        return self.connection.get("/jobs/{}".format(self.job_id)).json()
+        return self.connection.get("/jobs/{}".format(self.job_id), expected_status=200).json()
 
     def update_job(self, process_graph=None, output_format=None,
                    output_parameters=None, title=None, description=None,
@@ -51,33 +60,27 @@ class RESTJob(Job):
     def delete_job(self):
         """ Delete a job."""
         # DELETE /jobs/{job_id}
-        request = self.connection.delete("/jobs/{}".format(self.job_id))
-        assert request.status_code == 204
+        self.connection.delete("/jobs/{}".format(self.job_id), expected_status=204)
 
     def estimate_job(self):
         """ Calculate an time/cost estimate for a job."""
         # GET /jobs/{job_id}/estimate
-        return self.connection.get("/jobs/{}/estimate".format(self.job_id)).json()
+        return self.connection.get("/jobs/{}/estimate".format(self.job_id), expected_status=200).json()
 
     def start_job(self):
         """ Start / queue a job for processing."""
         # POST /jobs/{job_id}/results
-        url = "/jobs/{}/results".format(self.job_id)
-        request = self.connection.post(url)
-        if request.status_code != 202:
-            logger.warning("{u} returned with status code {s} instead of 202".format(u=url, s=request.status_code))
+        self.connection.post("/jobs/{}/results".format(self.job_id), expected_status=202)
 
     def stop_job(self):
         """ Stop / cancel job processing."""
         # DELETE /jobs/{job_id}/results
-        request = self.connection.delete("/jobs/{}/results".format(self.job_id))
+        self.connection.delete("/jobs/{}/results".format(self.job_id), expected_status=204)
 
-        return request.status_code
-
-    def list_results(self, type=None):
-        """ Get document with download links."""
-        # GET /jobs/{job_id}/results
-        raise NotImplementedError
+    @deprecated("Use :py:meth:`~RESTJOB.get_results` instead.", version="0.4.10")
+    def list_results(self) -> dict:
+        """Get batch job results metadata."""
+        return self.get_results().get_metadata()
 
     def download_result(self, target: Union[str, Path] = None) -> Path:
         """
@@ -87,11 +90,14 @@ class RESTJob(Job):
 
         :param target: String or path where the file should be downloaded to.
         """
-        return self.get_result().download_file(target)
+        return self.get_results().download_file(target=target)
 
+    @deprecated(
+        "Instead use :py:meth:`RESTJob.get_results` and the more flexible download functionality of :py:class:`JobResults`",
+        version="0.4.10")
     def download_results(self, target: Union[str, Path] = None) -> Dict[Path, dict]:
         """
-        Download job results into given folder (current working dir by default).
+        Download all job result files into given folder (current working dir by default).
 
         The names of the files are taken directly from the backend.
 
@@ -100,35 +106,27 @@ class RESTJob(Job):
         """
         return self.get_result().download_files(target)
 
+    @deprecated("Use :py:meth:`RESTJob.get_results` instead.", version="0.4.10")
     def get_result(self):
-        return Result(self)
+        return _Result(self)
 
-    @deprecated
-    def download(self, outputfile: str, outputformat=None):
-        """ Download the result as a raster."""
-        try:
-            return self.connection.download_job(self.job_id, outputfile, outputformat)
-        except ConnectionAbortedError as e:
-            return print(str(e))
+    def get_results(self) -> "JobResults":
+        """
+        Get handle to batch job results for result metadata inspection or downloading resulting assets.
+
+        .. versionadded:: 0.4.10
+        """
+        return JobResults(self)
 
     def status(self):
         """ Returns the status of the job."""
         return self.describe_job().get("status", "N/A")
 
-    @deprecated
-    def queue(self):
-        """ Queues the job. """
-        return self.connection.queue_job(self.job_id)
-
-    @deprecated
-    def results(self) -> List[RESTJobResult]:
-        """ Returns this job's results. """
-        return [RESTJobResult(link['href']) for link in self.connection.job_results(self.job_id)['links']]
-
     def logs(self, offset=None) -> List[JobLogEntry]:
         """ Retrieve job logs."""
-        return [JobLogEntry(log_entry['id'], log_entry['level'], log_entry['message'])
-                for log_entry in self.connection.job_logs(self.job_id, offset)['logs']]
+        url = "/jobs/{}/logs".format(self.job_id)
+        logs = self.connection.get(url, params={'offset': offset}, expected_status=200).json()["logs"]
+        return [JobLogEntry(log['id'], log['level'], log['message']) for log in logs]
 
     def run_synchronous(self, outputfile: Union[str, Path],
                         print=print, max_poll_interval=60, connection_retry_interval=30) -> 'RESTJob':
@@ -141,9 +139,10 @@ class RESTJob(Job):
             self.download_result(outputfile)
         return self
 
-    def start_and_wait(self, print=print, max_poll_interval: int = 60, connection_retry_interval: int = 30):
+    def start_and_wait(self, print=print, max_poll_interval: int = 60, connection_retry_interval: int = 30) -> "RESTJob":
         """
         Start the batch job, poll its status and wait till it finishes (or fails)
+
         :param print: print/logging function to show progress/status
         :param max_poll_interval: maximum number of seconds to sleep between status polls
         :param connection_retry_interval: how long to wait when status poll failed due to connection issue
@@ -184,3 +183,189 @@ class RESTJob(Job):
             ), job=self)
 
         return self
+
+
+class ResultAsset:
+    """
+    Result asset of a batch job (e.g. a GeoTIFF or JSON file)
+
+    .. versionadded:: 0.4.10
+    """
+
+    def __init__(self, job: RESTJob, name: str, href: str, metadata: dict):
+        self.job = job
+
+        self.name = name
+        """Asset name as advertised by the backend."""
+
+        self.href = href
+        """Download URL of the asset."""
+
+        self.metadata = metadata
+        """Asset metadata provided by the backend, possibly containing keys "type" (for media type), "roles", "title", "description"."""
+
+    def __repr__(self):
+        return "<ResultAsset {n!r} (type {t}) at {h!r}>".format(
+            n=self.name, t=self.metadata.get("type", "unknown"), h=self.href
+        )
+
+    def download(self, target: Optional[Union[Path, str]] = None, chunk_size=None) -> Path:
+        """
+        Download asset to given location
+
+        :param target: download target path. Can be an existing folder
+            (in which case the filename advertised by backend will be used)
+            or full file name. By default, the working directory will be used.
+        """
+        target = Path(target or Path.cwd())
+        if target.is_dir():
+            target = target / self.name
+        ensure_dir(target.parent)
+        logger.info("Downloading Job result asset {n!r} from {h!s} to {t!s}".format(n=self.name, h=self.href, t=target))
+        with target.open("wb") as f:
+            response = self._get_response(stream=True)
+            for block in response.iter_content(chunk_size=chunk_size):
+                f.write(block)
+        return target
+
+    def _get_response(self, stream=True) -> Response:
+        return self.job.connection.get(self.href, stream=stream)
+
+    def load_json(self) -> dict:
+        """Load asset in memory and parse as JSON."""
+        if not (self.name.lower().endswith(".json") or self.metadata.get("type") == "application/json"):
+            logger.warning("Asset might not be JSON")
+        return self._get_response().json()
+
+    def load_bytes(self) -> bytes:
+        """Load asset in memory as raw bytes."""
+        return self._get_response().content
+
+    # TODO: more `load` methods e.g.: load GTiff asset directly as numpy array
+
+
+class MultipleAssetException(OpenEoClientException):
+    pass
+
+
+class JobResults:
+    """
+    Results of a batch job: listing of one or more output files (assets)
+    and some metadata.
+
+    .. versionadded:: 0.4.10
+    """
+
+    def __init__(self, job: RESTJob):
+        self._job = job
+        self._results_url = "/jobs/{j}/results".format(j=self._job.job_id)
+        self._results = None
+
+    def __repr__(self):
+        return "<JobResults for job {j!r}>".format(j=self._job.job_id)
+
+    def get_metadata(self, force=False) -> dict:
+        """Get batch job results metadata (parsed JSON)"""
+        if self._results is None or force:
+            self._results = self._job.connection.get(self._results_url, expected_status=200).json()
+        return self._results
+
+    # TODO: provide methods for `stac_version`, `id`, `geometry`, `properties`, `links`, ...?
+
+    def get_assets(self) -> List[ResultAsset]:
+        """
+        Get all assets from the job results.
+        """
+        metadata = self.get_metadata()
+        if "assets" in metadata:
+            # API 1.0 style: dictionary mapping filenames to metadata dict (with at least a "href" field)
+            assets = metadata["assets"]
+        else:
+            # Best effort translation of on old style to "assets" style (#134)
+            assets = {a["href"].split("/")[-1]: a for a in metadata["links"]}
+        return [
+            ResultAsset(job=self._job, name=name, href=asset["href"], metadata=asset)
+            for name, asset in assets.items()
+        ]
+
+    def get_asset(self, name: str = None) -> ResultAsset:
+        """
+        Get single asset by name or without name if there is only one.
+        """
+        # TODO: also support getting a single asset by type or role?
+        assets = self.get_assets()
+        if len(assets) == 0:
+            raise OpenEoClientException("No assets in result.")
+        if name is None:
+            if len(assets) == 1:
+                return assets[0]
+            else:
+                raise MultipleAssetException("Multiple result assets for job {j}: {a}".format(
+                    j=self._job.job_id, a=[a.name for a in assets]
+                ))
+        else:
+            try:
+                return next(a for a in assets if a.name == name)
+            except StopIteration:
+                raise OpenEoClientException(
+                    "No asset {n!r} in: {a}".format(n=name, a=[a.name for a in assets])
+                )
+
+    def download_file(self, target: Union[Path, str] = None, name: str = None) -> Path:
+        """
+        Download single asset. Can be used when there is only one asset in the
+        :py:class:`JobResults`, or when the desired asset name is given explicitly.
+
+        :param target: path to download to. Can be an existing directory
+            (in which case the filename advertised by backend will be used)
+            or full file name. By default, the working directory will be used.
+        :param name: asset name to download (not required when there is only one asset)
+        :return: path of downloaded asset
+        """
+        try:
+            return self.get_asset(name=name).download(target=target)
+        except MultipleAssetException:
+            raise OpenEoClientException(
+                "Can not use `download_file` with multiple assets. Use `download_files` instead.")
+
+    def download_files(self, target: Union[Path, str] = None) -> List[Path]:
+        """
+        Download all assets to given folder.
+
+        :param target: path to folder to download to (must be a folder if it already exists)
+        :return: list of paths to the downloaded assets.
+        """
+        target = Path(target or Path.cwd())
+        if target.exists() and not target.is_dir():
+            raise OpenEoClientException("The target argument must be a folder. Got {t!r}".format(t=str(target)))
+        ensure_dir(target)
+        return [a.download(target) for a in self.get_assets()]
+
+
+@deprecated(reason="Use :py:class:`JobResults` instead", version="0.4.10")
+class _Result:
+    """
+    Wrapper around `JobResults` to adapt old deprecated "Result" API.
+
+    .. deprecated:: 0.4.10
+    """
+
+    # TODO: deprecated: remove this
+
+    def __init__(self, job):
+        self.results = JobResults(job=job)
+
+    def download_file(self, target: Union[str, Path] = None) -> Path:
+        return self.results.download_file(target=target)
+
+    def download_files(self, target: Union[str, Path] = None) -> Dict[Path, dict]:
+        target = Path(target or Path.cwd())
+        if target.exists() and not target.is_dir():
+            raise OpenEoClientException("The target argument must be a folder. Got {t!r}".format(t=str(target)))
+        return {a.download(target): a.metadata for a in self.results.get_assets()}
+
+    def load_json(self) -> dict:
+        return self.results.get_asset().load_json()
+
+    def load_bytes(self) -> bytes:
+        return self.results.get_asset().load_bytes()