+++ tests/test_batch_jobs.py
@@ -1,12 +1,9 @@
-import unittest
 from unittest import TestCase
 
 import requests_mock
-from mock import MagicMock
 
 import openeo
 from openeo.internal.graphbuilder_040 import GraphBuilder
-from openeo import Job
 from . import load_json_resource