+++ CHANGELOG.md
@@ -8,15 +8,32 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add "reflected" operator support to `ProcessBuilder` 
-
 ### Changed
-
-- Mark `Connection.imagecollection` more clearly as deprecated/legacy alias of `Connection.load_collection`
 
 ### Removed
 
 ### Fixed
+
+
+## [0.4.10] - 2021-02-26
+
+### Added
+
+- Add "reflected" operator support to `ProcessBuilder` 
+- Add `RESTJob.get_results()`, `JobResults` and `ResultAsset` for more fine-grained batch job result handling. (EP-3739)
+- Add documentation on batch job result (asset) handling and downloading
+
+### Changed
+
+- Mark `Connection.imagecollection` more clearly as deprecated/legacy alias of `Connection.load_collection`
+- Deprecated `job_results()` and `job_logs()` on `Connection` object, it's better to work through `RESTJob` object. 
+- Update `DataCube.sar_backscatter` to the latest process spec: add `coefficient` argument 
+  and remove `orthorectify`, `rtc`. ([openeo-processes#210](https://github.com/Open-EO/openeo-processes/pull/210))
+
+### Removed
+
+- Remove outdated batch job result download logic left-overs
+- Remove (outdated) abstract base class `openeo.job.Job`: did not add value, only caused maintenance overhead. ([#115](https://github.com/Open-EO/openeo-python-client/issues/115))
 
 
 ## [0.4.9] - 2021-01-29